widget compatible with wp58, fix 4 client issue
