+++ classes/class-s2-admin.php
@@ -1144,7 +1144,7 @@ class S2_Admin extends S2_Core {
 			return false;
 		}
 
-		if ( isset( $_POST['sub2-one-click-subscribe'] ) && 1 === $_POST['sub2-one-click-subscribe'] ) {
+		if ( isset( $_POST['sub2-one-click-subscribe'] ) && 1 === (int)$_POST['sub2-one-click-subscribe'] ) {
 			// Subscribe
 			$this->one_click_handler( $user_id, 'subscribe' );
 		} else {
