+++ classes/class-s2-core.php
@@ -1,5 +1,8 @@
 <?php
+require_once S2PATH . 'traits/ShortcodeTrait.php';
+
 class S2_Core {
+	use Shortcode;
 	// variables and constructor are declared at the end
 	/**
 	 * Load translations
@@ -1393,7 +1396,9 @@ class S2_Core {
 	 */
 	public function subscribe2_widget() {
 		require_once S2PATH . 'classes/class-s2-form-widget.php';
+		require_once S2PATH . 'admin/WidgetPreview.php';
 		register_widget( 'S2_Form_Widget' );
+		// new WidgetPreview();
 	}
 
 	/**
@@ -2041,7 +2046,6 @@ class S2_Core {
 			}
 
 			// add the frontend filters
-			add_shortcode( 'subscribe2', array( &$this, 'shortcode' ) );
 			add_filter( 'the_content', array( &$this, 'filter' ), 10 );
 
 			// add actions for other plugins
@@ -2055,6 +2059,7 @@ class S2_Core {
 				add_action( 'wp_footer', array( &$this, 'js_ip_library_script' ), 20 );
 			}
 		}
+		add_shortcode( 'subscribe2', array( &$this, 'widget_shortcode' ) );
 	}
 
 	/* ===== define some variables ===== */
